# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tutor Marketplace CRM & Lead Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Staff login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with pipeline info
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /leads/{id}/status:
    patch:
      operationId: updateLeadStatus
      tags: [leads]
      summary: Update lead status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadStatusUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/assign:
    patch:
      operationId: assignLead
      tags: [leads]
      summary: Assign lead to staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadAssign"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  # Tutors
...
  /requirements/{id}/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications for a requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Tutor applies for a requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # Demos
  /demos:
    get:
      operationId: listDemos
      tags: [demos]
      summary: List demos
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Demos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Demo"
    post:
      operationId: createDemo
      tags: [demos]
      summary: Schedule a demo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoInput"
      responses:
        "201":
          description: Demo scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demo"
  /demos/{id}:
    get:
      operationId: getDemo
      tags: [demos]
      summary: Get demo details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Demo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demo"
    patch:
      operationId: updateDemo
      tags: [demos]
      summary: Update demo status/feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoUpdate"
      responses:
        "200":
          description: Updated demo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demo"

  # Payments
...
  /staff:
    get:
      operationId: listStaff
      tags: [auth]
      summary: List staff users
      responses:
        "200":
          description: Staff users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffUser"
    post:
      operationId: createStaff
      tags: [auth]
      summary: Create staff user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/StaffUser"
    StaffUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, parent_success_executive, tutor_acquisition_executive]
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, parent_success_executive, tutor_acquisition_executive]
...
    Lead:
      type: object
      required: [id, parentId, parentName, parentPhone, area, status, source, createdAt]
      properties:
        id:
          type: integer
        parentId:
          type: integer
        parentName:
          type: string
        parentPhone:
          type: string
        area:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, demo_scheduled, demo_done, converted, lost]
        source:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assignedName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, contacted, qualified, demo_scheduled, demo_done, converted, lost]
        notes:
          type: string
    LeadAssign:
      type: object
      required: [staffId]
      properties:
        staffId:
          type: integer
...
    Application:
      type: object
      required: [id, requirementId, tutorId, status, createdAt]
      properties:
        id:
          type: integer
        requirementId:
          type: integer
        tutorId:
          type: integer
        tutorName:
          type: string
        tutorPhone:
          type: string
        status:
          type: string
          enum: [pending, shortlisted, rejected, selected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ApplicationInput:
      type: object
      required: [tutorId]
      properties:
        tutorId:
          type: integer
        notes:
          type: string
    ApplicationUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, shortlisted, rejected, selected]
        notes:
          type: string
    Demo:
      type: object
      required: [id, requirementId, tutorId, scheduledAt, status, createdAt]
      properties:
        id:
          type: integer
        requirementId:
          type: integer
        tutorId:
          type: integer
        tutorName:
          type: string
        parentId:
          type: integer
        parentName:
          type: string
        parentPhone:
          type: string
        subject:
          type: string
        area:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
        feedback:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        createdAt:
          type: string
    DemoInput:
      type: object
      required: [requirementId, tutorId, scheduledAt]
      properties:
        requirementId:
          type: integer
        tutorId:
          type: integer
        scheduledAt:
          type: string
        assignedTo:
          type: integer
        notes:
          type: string
    DemoUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
        feedback:
          type: string
        outcome:
          type: string
        assignedTo:
          type: integer
...