# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tutor Marketplace CRM & Lead Management API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffInputRole } from './staffInputRole';

export interface StaffInput {
  name: string;
  email: string;
  password: string;
  role: StaffInputRole;
}