# lib/api-zod/src/generated/types/staffUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tutor Marketplace CRM & Lead Management API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffUserRole } from './staffUserRole';

export interface StaffUser {
  id: number;
  name: string;
  email: string;
  role: StaffUserRole;
  createdAt?: string;
}